Catch failure to SHOW CREATE TABLE.
git-svn-id: http://maatkit.googlecode.com/svn/trunk@4160 dfb901c2-3250-0410-b216-0b33211c9131

+++ common/MySQLDump.pm
@@ -64,6 +64,7 @@ sub dump {
 
    if ( $what eq 'table' ) {
       my $ddl = $self->get_create_table($dbh, $quoter, $db, $tbl);
+      return unless $ddl;
       if ( $ddl->[0] eq 'table' ) {
          return $before
             . 'DROP TABLE IF EXISTS ' . $quoter->quote($tbl) . ";\n"
@@ -148,7 +149,12 @@ sub get_create_table {
       my $curr_db = $self->_use_db($dbh, $quoter, $db);
       $sql = "SHOW CREATE TABLE " . $quoter->quote($db, $tbl);
       MKDEBUG && _d($sql);
-      my $href = $dbh->selectrow_hashref($sql);
+      my $href;
+      eval { $href = $dbh->selectrow_hashref($sql); };
+      if ( $EVAL_ERROR ) {
+         warn "Failed to $sql.  The table may be damaged.\nError: $EVAL_ERROR";
+         return;
+      }
       $self->_use_db($dbh, $quoter, $curr_db);
       $sql = '/*!40101 SET @@SQL_MODE := @OLD_SQL_MODE, '
          . '@@SQL_QUOTE_SHOW_CREATE := @OLD_QUOTE */';
